Move base ValidationError to new module base_exceptions
